agora se pode logar no site usando uma conta Google / folha de estilo adicionada a pagina Home
