+++ quinzena7/labefy-project/package.json
@@ -9,9 +9,11 @@
     "axios": "^0.21.1",
     "react": "^17.0.2",
     "react-dom": "^17.0.2",
+    "react-google-login": "^5.2.2",
     "react-player": "^2.9.0",
     "react-scripts": "4.0.3",
     "styled-components": "^5.3.0",
+    "surge": "^0.23.0",
     "web-vitals": "^1.1.2"
   },
   "scripts": {
