+++ quinzena7/labefy-project/src/App.css
@@ -1,0 +1,15 @@
+@import url('https://fonts.googleapis.com/css?family=Poppins:200,300,400,500,600,700,800,900&display=swap');
+
+* {
+    box-sizing: border-box;
+    margin: 0;
+    padding: 0;
+    font-family: 'Poppins', sans-serif;
+}
+
+.google-button {
+    position: absolute;
+    left: 85%;
+    top: 8%;
+    z-index: 1001;
+}
